+++ src/test/java/roomescape/auth/controller/LoginControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
